Hint for noobs
Hi David,
first of all thanks for your commitment to handle ebay apis.
I was trying to get the user token as mentioned, but copied the code from the browser to test the function ``getUserToken``. I completly forgot to urldecode the code. Maybe a little hint for this procedure would help other guys too.